# Optimistically render first 10 comments and lazily render remaining comments