Initial render of comments would be perceived by the user as 3.4x quicker from 1762ms to 512ms on a 4x slow down of CPU.

![Screen Shot 2021-03-16 at 17 40 52](https://user-images.githubusercontent.com/638051/111355750-94c77400-867f-11eb-9b82-b5c875b19d08.png)

(10 refreshes and clicking on the button, x-axis showing value for each run)

### Before - Full render is slow
![Screen Shot 2021-03-16 at 15 31 35](https://user-images.githubusercontent.com/638051/111355775-9abd5500-867f-11eb-9d7a-8b624bd0c10f.png)

### After - Initial rendering early with full rendering not heavily affected
![Screen Shot 2021-03-16 at 15 46 28](https://user-images.githubusercontent.com/638051/111355766-998c2800-867f-11eb-8ba7-1ed918be6cda.png)

## Part rendered with loading state


|Before |After |
| ----------- |----------- |
|![2021-03-16 17 52 35](https://user-images.githubusercontent.com/638051/111356755-96456c00-8680-11eb-9be8-82d5e3647c4f.gif)| ![2021-03-16 16 10 12](https://user-images.githubusercontent.com/638051/111355784-9c871880-867f-11eb-839b-f874e0a39eb8.gif)    |